+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADEntitlementManagementAccessPackageAssignmentPolicy/MSFT_AADEntitlementManagementAccessPackageAssignmentPolicy.psm1
@@ -103,10 +103,13 @@ function Get-TargetResource
         $nullResult.Ensure = 'Absent'
 
         $getValue = $null
-        $getValue = Get-MgBetaEntitlementManagementAccessPackageAssignmentPolicy `
-            -AccessPackageAssignmentPolicyId $id `
-            -ExpandProperty "customExtensionHandlers(`$expand=customExtension)" `
-            -ErrorAction SilentlyContinue
+        if (-not [System.String]::IsNullOrEmpty($id))
+        {
+            $getValue = Get-MgBetaEntitlementManagementAccessPackageAssignmentPolicy `
+                -AccessPackageAssignmentPolicyId $id `
+                -ExpandProperty "customExtensionHandlers(`$expand=customExtension)" `
+                -ErrorAction SilentlyContinue
+        }
 
         if ($null -eq $getValue)
         {
@@ -288,9 +291,18 @@ function Get-TargetResource
         }
         #endregion
 
+        $AccessPackageIdValue = $getValue.AccessPackageId
+        $ObjectGuid = [System.Guid]::empty
+        $isGUID = [System.Guid]::TryParse($AccessPackageIdValue, [System.Management.Automation.PSReference]$ObjectGuid)
+        if ($isGUID)
+        {
+            $accesspackage = Get-MgBetaEntitlementManagementAccessPackage -AccessPackageId $AccessPackageIdValue
+            $AccessPackageIdValue = $accesspackage.DisplayName
+        }
+
         $results = @{
             Id                      = $getValue.Id
-            AccessPackageId         = $getValue.AccessPackageId
+            AccessPackageId         = $AccessPackageIdValue
             AccessReviewSettings    = $formattedAccessReviewSettings
             CanExtend               = $getValue.CanExtend
             CustomExtensionHandlers = $formattedCustomExtensionHandlers
@@ -528,6 +540,30 @@ function Set-TargetResource
             }
             $CreateParameters.CustomExtensionHandlers = $formattedCustomExtensionHandlers
         }
+
+        # Check to see if the AccessPackageId is in GUID form. If not, resolve it by name.
+        if (-not [System.String]::IsNullOrEmpty($AccessPackageId))
+        {
+            $ObjectGuid = [System.Guid]::empty
+            $isGUID = [System.Guid]::TryParse($AccessPackageId, [System.Management.Automation.PSReference]$ObjectGuid)
+            if (-not $isGUID)
+            {
+                # Retrieve by name
+                Write-Verbose -Message "Retrieving Entitlement Management Access Package by Name {$AccessPackageId}"
+                $package = Get-MgBetaEntitlementManagementAccessPackage -Filter "displayName eq '$AccessPackageId'"
+                if ($null -ne $package)
+                {
+                    $AccessPackageId = $package.Id
+                }
+                else
+                {
+                    throw "Could not retrieve the Access Package using identifier {$AccessPackageId}"
+                }
+            }
+            $CreateParameters.AccessPackageId = $AccessPackageId
+        }
+
+        Write-Verbose -Message "Creating with Values: $(Convert-M365DscHashtableToString -Hashtable $CreateParameters)"
         New-MgBetaEntitlementManagementAccessPackageAssignmentPolicy `
             -BodyParameter $CreateParameters
     }
@@ -624,6 +660,27 @@ function Set-TargetResource
                 }
             }
             $UpdateParameters.CustomExtensionHandlers = $formattedCustomExtensionHandlers
+        }
+
+        if (-not [System.String]::IsNullOrEmpty($AccessPackageId))
+        {
+            $ObjectGuid = [System.Guid]::empty
+            $isGUID = [System.Guid]::TryParse($AccessPackageId, [System.Management.Automation.PSReference]$ObjectGuid)
+            if (-not $isGUID)
+            {
+                # Retrieve by name
+                Write-Verbose -Message "Retrieving Entitlement Management Access Package by Name {$AccessPackageId}"
+                $package = Get-MgBetaEntitlementManagementAccessPackage -Filter "displayName eq '$AccessPackageId'"
+                if ($null -ne $package)
+                {
+                    $AccessPackageId = $package.Id
+                }
+                else
+                {
+                    throw "Could not retrieve the Access Package using identifier {$AccessPackageId}"
+                }
+            }
+            $UpdateParameters.AccessPackageId = $AccessPackageId
         }
 
         #write-verbose ($UpdateParameters|convertto-json -Depth 100)