+++ Modules/Microsoft365DSC/Examples/Resources/AADAuthenticationMethodPolicyAuthenticator/2-Update.ps1
@@ -21,20 +21,12 @@ Configuration Example
             Ensure                = "Present";
             ExcludeTargets        = @(
                 MSFT_AADAuthenticationMethodPolicyAuthenticatorExcludeTarget{
-                    Id = 'Legal Team'
-                    TargetType = 'group'
-                }
-                MSFT_AADAuthenticationMethodPolicyAuthenticatorExcludeTarget{
-                    Id = 'Paralegals'
+                    Id = 'Finance Team' # Updated Property
                     TargetType = 'group'
                 }
             );
             FeatureSettings       = MSFT_MicrosoftGraphmicrosoftAuthenticatorFeatureSettings{
                 DisplayLocationInformationRequiredState = MSFT_MicrosoftGraphAuthenticationMethodFeatureConfiguration{
-                    ExcludeTarget = MSFT_AADAuthenticationMethodPolicyAuthenticatorFeatureTarget{
-                        Id = 'all_users'
-                        TargetType = 'group'
-                    }
                     IncludeTarget = MSFT_AADAuthenticationMethodPolicyAuthenticatorFeatureTarget{
                         Id = 'all_users'
                         TargetType = 'group'
@@ -42,10 +34,6 @@ Configuration Example
                     State = 'default'
                 }
                 CompanionAppAllowedState = MSFT_MicrosoftGraphAuthenticationMethodFeatureConfiguration{
-                    ExcludeTarget = MSFT_AADAuthenticationMethodPolicyAuthenticatorFeatureTarget{
-                        Id = 'all_users'
-                        TargetType = 'group'
-                    }
                     IncludeTarget = MSFT_AADAuthenticationMethodPolicyAuthenticatorFeatureTarget{
                         Id = 'all_users'
                         TargetType = 'group'
@@ -53,10 +41,6 @@ Configuration Example
                     State = 'default'
                 }
                 DisplayAppInformationRequiredState = MSFT_MicrosoftGraphAuthenticationMethodFeatureConfiguration{
-                    ExcludeTarget = MSFT_AADAuthenticationMethodPolicyAuthenticatorFeatureTarget{
-                        Id = 'all_users'
-                        TargetType = 'group'
-                    }
                     IncludeTarget = MSFT_AADAuthenticationMethodPolicyAuthenticatorFeatureTarget{
                         Id = 'all_users'
                         TargetType = 'group'
@@ -75,7 +59,7 @@ Configuration Example
                     TargetType = 'group'
                 }
             );
-            IsSoftwareOathEnabled = $True; # Updated Property
+            IsSoftwareOathEnabled = $False;
             State                 = "enabled";
         }
     }