+++ Modules/Microsoft365DSC/Examples/Resources/AADAuthenticationMethodPolicyX509/2-Update.ps1
@@ -18,9 +18,6 @@ Configuration Example
         AADAuthenticationMethodPolicyX509 "AADAuthenticationMethodPolicyX509-X509Certificate"
         {
             AuthenticationModeConfiguration = MSFT_MicrosoftGraphx509CertificateAuthenticationModeConfiguration{
-
-                Rules = @(@()
-                )
                 X509CertificateAuthenticationDefaultMode = 'x509CertificateSingleFactor'
             };
             CertificateUserBindings         = @(
@@ -55,7 +52,7 @@ Configuration Example
                     TargetType = 'group'
                 }
             );
-            State                           = "disabled"; # Updated Property
+            State                           = "enabled";
         }
     }
 }